requirements.txt
numpy >= 1.25.2, <= 1.26.4
rdkit >= 2023.9.6
openbabel-wheel == 3.1.1.22
networkx >= 3.0, <= 3.3
matplotlib >= 3.10.0
